Extract eval and factory functions out. Remove addSourceUrl from production modules.
