+++ lib/amd/captureDefines.js
@@ -50,8 +50,6 @@ function captureDefines (source) {
 	// indicate we are AMD and we can handle the jqueries
 	capture.amd = { jQuery: {} };
 
-	// Note: V8 intermittently fails if we embed eval() in new Function()
-	// and source has "use strict" in it
 	amdEval(global, capture, source);
 
 	if (!result) {
