+++ lib/node/factory.js
@@ -1,15 +1,12 @@
 /** @license MIT License (c) copyright 2014 original authors */
 /** @author Brian Cavalier */
 /** @author John Hann */
+
+var es5Transform = require('../es5Transform');
+var createRequire = require('../createRequire');
+var nodeEval = require('./eval');
+
 module.exports = nodeFactory;
-
-var es5Transform = require('./es5Transform');
-var createRequire = require('./createRequire');
-var nodeEval = require('./node/eval');
-
-var _global;
-
-_global = typeof global !== 'undefined' ? global : window;
 
 function nodeFactory (loader, load) {
 	var name, source, exports, module, require;
@@ -21,8 +18,7 @@ function nodeFactory (loader, load) {
 	require = createRequire(loader, name);
 
 	return function () {
-		// TODO: use loader.global when es6-module-loader implements it
-		nodeEval(_global, require, exports, module, source);
+		nodeEval(global, require, exports, module, source);
 		// figure out what author intended to export
 		return exports === module.exports
 			? exports // a set of named exports
