+++ lib/script/factory.js
@@ -2,12 +2,12 @@
 /** @author Brian Cavalier */
 /** @author John Hann */
 
-var globalEval = require('./script/eval');
+var scriptEval = require('./eval');
 
-module.exports = globalFactory;
+module.exports = scriptFactory;
 
-function globalFactory (loader, load) {
+function scriptFactory (loader, load) {
 	return function () {
-		globalEval(load.source);
+		scriptEval(load.source);
 	};
 }
